using package json and npm link. this allows mockery to work as expected, but tests must be run in serial. Had to move files around to create `foo-lib` module.
