+++ foo/index.js
