+++ foo/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "foo-lib",
+  "version": "1.0.0",
+  "description": "foo lib",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "foo"
+  },
+  "author": "",
+  "license": "ISC"
+}
